--- pages/test.tsx
@@ -1,1 +1,0 @@
-export { default } from "containers/test";
