+++ src/containers/landing/components/section/MainBanner.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import DOMPurify from "isomorphic-dompurify";
 import media from "styles/media";
-import Store from "../common/store/Store";
+import Store from "components/common/store/Store";
 import { BANNER_CONTENTS } from "database/main";
 
 const MainBanner = () => {
