+++ src/containers/landing/components/section/SecondSection.tsx
