+++ src/containers/landing/components/section/ThirdSection.tsx
