--- src/containers/test/index.tsx
@@ -1,4 +1,0 @@
-import React from "react";
-export default function TestPage() {
-  return <div></div>;
-}
